+++ src/lib/charset-utils.ts
@@ -1,3 +1,5 @@
+import { getCharCount } from "@trt2/gsm-charset-utils";
+
 const gsmReplacements = [
   ["‘", "'"],
   ["’", "'"],
@@ -12,6 +14,9 @@ export const replaceEasyGsmWins = (text: string) =>
     (acc, replacement) => acc.replace(replacement[0], replacement[1]),
     text
   );
+
+export const getSpokeCharCount = (text: string) =>
+  getCharCount(replaceEasyGsmWins(text));
 
 export const replaceCurlyApostrophes = (rawText: string) =>
   rawText.replace(/[\u2018\u2019]/g, "'");